+++ crates/cubecl-runtime/src/channel/cell.rs
@@ -60,7 +60,7 @@ where
         self.server.borrow_mut().create(resource)
     }
 
-    fn empty(&self, size: usize) -> Handle {
+    fn empty(&self, size: u64) -> Handle {
         self.server.borrow_mut().empty(size)
     }
